Update: Song Select Screen

+++ Assets/Scripts/Quaver.Graphics/BackgroundBufferer.cs
@@ -58,7 +58,7 @@ namespace Quaver.Graphics
                     _bufferSprite.GetComponent<SpriteRenderer>().sprite = sprite;
                     _bufferSprite.GetComponent<BackgroundDimAnimator>().dim = _bufferSprite.GetComponent<BackgroundDimAnimator>().Manager.SelectScreenDim;
 
-                    if (sprite.rect.size.y / sprite.rect.size.x <= Screen.height / Screen.width)
+                    if (sprite.rect.size.y / sprite.rect.size.x <= (float)Screen.height / (float)Screen.width)
                         _bufferSprite.transform.localScale = Vector3.one * 1.05f * (20f * (100f / (float)sprite.rect.size.y));
                     else
                         _bufferSprite.transform.localScale = Vector3.one * 1.05f * ((float)Screen.width / (float)Screen.height) * 20f * (100f / (float)sprite.rect.size.x);
